% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfethemes-package.R, R/ggdfe-package.R
\docType{package}
\name{ggdfe-package}
\alias{ggdfe}
\alias{ggdfe-package}
\title{ggdfe: A Theme for ggplot in Department for Education House Style}
\description{
Apply the house style of the Department for Education to ggplot, including colours and themes, for consistent reproducable charting for R users at DfE.

Apply the house style of the Department for Education to ggplot, including colours and themes, for consistent reproducable charting for R users at DfE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dfe-r-community.github.io/ggdfe/}
}


Useful links:
\itemize{
  \item \url{https://dfe-r-community.github.io/ggdfe/}
}

}
\author{
\strong{Maintainer}: Sam Cooke \email{sam.cooke@education.gov.uk}

}
\keyword{internal}
