% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dfe_dark.R, R/theme_dfe_light.R,
%   R/theme_nimbus.R
\name{themes}
\alias{themes}
\alias{theme_dfe_dark}
\alias{theme_dfe_light}
\alias{theme_nimbus}
\title{theme_dfe_dark}
\usage{
theme_dfe_dark()

theme_dfe_light(base_axis = "x")

theme_nimbus(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  all_gridlines = FALSE
)
}
\arguments{
\item{base_axis}{sets the orientation of the base and grid lines (defaults \code{x})}

\item{base_size, base_family, base_line_size, base_rect_size}{See
\code{ggplot2::theme_grey()} for details}

\item{all_gridlines}{Set \code{TRUE} to show all panel gridlines - by default,
the vertical gridlines will not be shown.}
}
\value{
A dark ggplot2 DfE theme

A light ggplot2 DfE theme

Use \code{base_axis} to switch between horizontal and vertical gridlines.
Use \code{y} for horizontal bar charts or \code{both} for all gridlines

A ggplot2 plot theme
}
\description{
theme_dfe_dark

theme_dfe_light

DfE Plot Themes
}
\examples{
library(ggplot2)

plot <- ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width, colour = Species))+
geom_jitter()+
scale_colour_dfe(palette = 'likert')+
  labs(title = "This is the title",
         subtitle = "This is a subtitle",
                caption = "This is a caption | Source: Iris dataset")
plot + theme_dfe_dark()

library(ggplot2)

plot <- ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width, colour = Species))+
geom_jitter()+
scale_colour_dfe()+
  labs(title = "This is the title",
         subtitle = "This is a subtitle",
                caption = "This is a caption | Source: Iris dataset")

plot + theme_dfe_light()

plot <- ggplot(data = mtcars, aes(x=cyl, y=hp, fill = cyl))+
  geom_col()+
  coord_flip()+
  theme_dfe_light(base_axis = "y")

plot

library(ggplot2)

p1 <- ggplot(diamonds, aes(carat, price)) + geom_point()
p2 <- p1 + facet_wrap(~cut)

p1 + theme_nimbus() + ggtitle("theme_nimbus()")
p2 + theme_nimbus() + ggtitle("theme_nimbus()")
}
