% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nimbus.R
\name{themes}
\alias{themes}
\alias{theme_nimbus}
\title{DfE Plot Themes}
\usage{
theme_nimbus(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  all_gridlines = FALSE
)
}
\arguments{
\item{base_size, base_family, base_line_size, base_rect_size}{See
\code{ggplot2::theme_grey()} for details}

\item{all_gridlines}{Set \code{TRUE} to show all panel gridlines - by default,
the vertical gridlines will not be shown.}
}
\value{
A ggplot2 plot theme
}
\description{
DfE Plot Themes
}
\examples{
library(ggplot2)

p1 <- ggplot(diamonds, aes(carat, price)) + geom_point()
p2 <- p1 + facet_wrap(~cut)

p1 + theme_nimbus() + ggtitle("theme_nimbus()")
p2 + theme_nimbus() + ggtitle("theme_nimbus()")
}
